refactor: extract player specific logic to `player.c`
