+++ src/ecs/entities.c
@@ -12,8 +12,6 @@
 #define GET_COMPONENT(mValue, mEntity) SCENE_GET_COMPONENT_PTR(params->scene, mEntity, mValue)
 #define ENTITY_HAS_DEPS(mEntity, mDependencies) ((params->scene->components.tags[mEntity] & (mDependencies)) == (mDependencies))
 
-// TODO(thismarvin): Ideally the following would be in something like `src/entities/player.c`.
-
 static void ApplyResolutionPerfectly
 (
     CPosition* position,
@@ -39,107 +37,6 @@ static void ApplyResolutionPerfectly
     {
         position->value.y = RectangleBottom(otherAabb);
     }
-}
-
-static bool PlayerIsVulnerable(const CPlayer* player)
-{
-    return player->invulnerableTimer >= player->invulnerableDuration;
-}
-
-static void PlayerOnDamage(const OnDamageParams* params)
-{
-    assert(ENTITY_HAS_DEPS(params->entity, TAG_PLAYER | TAG_MORTAL));
-
-    CPlayer* player = GET_COMPONENT(player, params->entity);
-    CMortal* mortal = GET_COMPONENT(mortal, params->entity);
-
-    const CDamage* otherDamage = GET_COMPONENT(otherDamage, params->otherEntity);
-
-    if (!PlayerIsVulnerable(player))
-    {
-        return;
-    }
-
-    mortal->hp -= otherDamage->value;
-    player->invulnerableTimer = 0;
-}
-
-static OnCollisionResult PlayerOnCollision(const OnCollisionParams* params)
-{
-    assert(ENTITY_HAS_DEPS(params->entity, TAG_PLAYER | TAG_POSITION | TAG_KINETIC | TAG_MORTAL));
-
-    CPlayer* player = GET_COMPONENT(player, params->entity);
-    CPosition* position = GET_COMPONENT(position, params->entity);
-    CKinetic* kinetic = GET_COMPONENT(kinetic, params->entity);
-    const CMortal* mortal = GET_COMPONENT(mortal, params->entity);
-
-    // Collision specific logic that will not resolve the player.
-    {
-        if (ENTITY_HAS_DEPS(params->otherEntity, TAG_WALKER | TAG_DAMAGE))
-        {
-            OnDamageParams onDamageParams = (OnDamageParams)
-            {
-                .scene = params->scene,
-                .entity = params->entity,
-                .otherEntity = params->otherEntity,
-            };
-
-            mortal->onDamage(&onDamageParams);
-
-            return ON_COLLISION_RESULT_NONE;
-        }
-    }
-
-    // Collision leeway.
-    {
-        // Check if the player hit its head on the bottom of a collider.
-        if (params->resolution.y > 0 && fabsf(params->overlap.width) <= 4)
-        {
-            if (params->aabb.x < params->otherAabb.x)
-            {
-                position->value.x = RectangleLeft(params->otherAabb) - params->aabb.width;
-            }
-            else
-            {
-                position->value.x = RectangleRight(params->otherAabb);
-            }
-
-            return (OnCollisionResult)
-            {
-                .xAxisResolved = true,
-                .yAxisResolved = false,
-            };
-        }
-    }
-
-    // Resolve collision.
-    ApplyResolutionPerfectly(position, params->aabb, params->otherAabb, params->resolution);
-
-    // Resolution specific player logic.
-    {
-        if ((params->resolution.x < 0 && kinetic->velocity.x > 0)
-                || (params->resolution.x > 0 && kinetic->velocity.x < 0))
-        {
-            kinetic->velocity.x = 0;
-        }
-
-        if ((params->resolution.y < 0 && kinetic->velocity.y > 0)
-                || (params->resolution.y > 0 && kinetic->velocity.y < 0))
-        {
-            kinetic->velocity.y = 0;
-        }
-
-        if (params->resolution.y < 0)
-        {
-            player->grounded = true;
-        }
-    }
-
-    return (OnCollisionResult)
-    {
-        .xAxisResolved = params->resolution.x != 0,
-        .yAxisResolved = params->resolution.y != 0,
-    };
 }
 
 // TODO(thismarvin): The collider no longer scales...
@@ -194,104 +91,6 @@ static OnCollisionResult WalkerOnCollision(const OnCollisionParams* params)
     {
         .xAxisResolved = params->resolution.x != 0,
         .yAxisResolved = params->resolution.y != 0,
-    };
-}
-
-EntityBuilder ECreatePlayer(const f32 x, const f32 y)
-{
-    Deque components = DEQUE_OF(Component);
-
-    const u64 tags =
-        TAG_NONE
-        | TAG_POSITION
-        | TAG_DIMENSION
-        | TAG_COLOR
-        | TAG_SPRITE
-        | TAG_KINETIC
-        | TAG_SMOOTH
-        | TAG_COLLIDER
-        | TAG_PLAYER
-        | TAG_MORTAL;
-
-    Vector2 position = Vector2Create(x, y);
-
-    ADD_COMPONENT(CPosition, ((CPosition)
-    {
-        .value = position,
-    }));
-
-    ADD_COMPONENT(CDimension, ((CDimension)
-    {
-        .width = 15,
-        .height = 35,
-    }));
-
-    ADD_COMPONENT(CColor, ((CColor)
-    {
-        .value = COLOR_WHITE,
-    }));
-
-    ADD_COMPONENT(CSprite, ((CSprite)
-    {
-        .source = (Rectangle) { 16, 0, 32, 48 },
-        .offset = Vector2Create(-8, -13),
-    }));
-
-    ADD_COMPONENT(CKinetic, ((CKinetic)
-    {
-        .velocity = VECTOR2_ZERO,
-        .acceleration = Vector2Create(0, 1000),
-    }));
-
-    ADD_COMPONENT(CSmooth, ((CSmooth)
-    {
-        .previous = position,
-    }));
-
-    ADD_COMPONENT(CCollider, ((CCollider)
-    {
-        .layer = LAYER_NONE,
-        .mask = LAYER_ALL,
-        .onCollision = PlayerOnCollision,
-    }));
-
-    ADD_COMPONENT(CMortal, ((CMortal)
-    {
-        .hp = 2,
-        .onDamage = PlayerOnDamage,
-    }));
-
-    f32 jumpHeight = 16 * 3 + 6;
-    f32 jumpDuration = 0.4;
-
-    f32 jumpGravity = (2 * jumpHeight) / (jumpDuration * jumpDuration);
-    f32 defaultGravity = jumpGravity * 1.5;
-
-    f32 jumpVelocity = jumpGravity * jumpDuration;
-
-    f32 coyoteDuration = CTX_DT * 6;
-    f32 invulnerableDuration = 1.5f;
-
-    ADD_COMPONENT(CPlayer, ((CPlayer)
-    {
-        .groundedLastFrame = false,
-        .grounded = false,
-        .coyoteTimer = coyoteDuration,
-        .coyoteDuration = coyoteDuration,
-        .moveSpeed = 200,
-        .jumping = false,
-        .dead = false,
-        .jumpVelocity = jumpVelocity,
-        .jumpGravity = jumpGravity,
-        .defaultGravity = defaultGravity,
-        .invulnerableTimer = invulnerableDuration,
-        .invulnerableDuration = invulnerableDuration,
-    }));
-
-    return (EntityBuilder)
-    {
-        .tags = tags,
-        .components = components,
     };
 }
